+++ src/main/java/com/alexanders/recipeapp/repositories/RecipeRepository.java
@@ -3,5 +3,5 @@ package com.alexanders.recipeapp.repositories;
 import com.alexanders.recipeapp.domain.Recipe;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RecipeRepository extends CrudRepository<Recipe, Long> {
+public interface RecipeRepository extends CrudRepository<Recipe, Integer> {
 }
